move testmonial into indexpage
